shore: Move endpoint implementations to a separate module so server/mod.rs file do not gets too complex
