+++ src/server/endpoints.rs
@@ -1,0 +1,3 @@
+pub mod account;
+pub mod block;
+pub mod transaction;
